Fix duplicate import and remove font fetch
